Set logic from main.js to presenter/trip.js
